Fix adaptive icon on "round" launchers
We use a android:roundIcon for bckwards compatibility with android < 8.
This caused the adaptive icon to not be used on round icon launchers.
This branch just copies the regular adaptive icon xml to a round one.

Once we don't support android < 8 any more we can completely drop
the roundIcon stuff as the normal adaptive icon works just fine
for round launchers unless you use android:roundIcon in the xml.