Fix encryption flag erraneously set for "ws" connections
This could cause the client to think that there is a certificate
mismatch if there is a pinned certificate for this host already
and we then connect to an unecrypted ws socket.

Also, disconnect clients before raising the certificate mismatch
error in case a client implementation would just ignore this
and reconnect syncronously right way.